close search:fix: hide results when focusing on language picker
as blur event hides results to early when clicking on a result,
which means its not going to the result link...

this hack does the job. but it does not close results when pressing
shift+tab from the search field. i dont think it matters much, since only the the
inventaire logo is clickable (and a skip link probably later).