show less/more details
based on overflow instead of a selection of important claims (elsewhere
called "shorlist"), as its not the epriority for the moment, but welcome
to do it later.

tweak WrapToggler to initialize with less info and expand on
click. There might be a smarter and shorter way to do, but hey this
works