work and edition layout: add infobox claims
edition claims are not definitive and inspired by small editors websites
([libertalia](https://www.editionslibertalia.com/catalogue/ceux-d-en-bas/),
[la fabrique](https://lafabrique.fr/les-patrons-de-la-presse-nationale/),
[communs](https://www.editionsducommun.org/collections/all/products/se-faire-virer-manon-delatre)
to name a few)
which have a pattern of ordering information of an edition as follow:
- date of publication
- number of pages
- isbn 13

This commit also tweaks WrapToggler to initialize with less info and expand on
click. There should be a smarter and shorter way to do, but couldnt find it yet.

[suggestions]
- add a "by" before authors line to emphasis that this are authors
- rename "edition of" claim key -> "other editions"
(as "edition of" may be cryptic for people who dont want to make the difference
between work and edition), "other editions" as been seen on babelio edition page,
next to publisher display