move loading message within selector's dropdown
Since it can take some time to load facets, and information on what to select
can be displayed as soon as advanced option is clicked on.

To do this, we extracted the sync part out of `getWorksFacets`, and allow
selector to have a default value placeholder

A friendly message is displayed within the selector (reviewer: don't be lured
by the absence of i18n key, CSS should be correct when key exists)

This is preparing ground for adding large reverse claims values to get filtered
(ie. wdt:P69 educated at, or wdt:P166 award received)