scss: prevent max-width media queries to include the thresold value
to avoid having both rules below and above the threshold to apply at the same time.

max-width and min-width are inclusive comparators (<= and >=), there is unforunately
no strict comparators (< and >). An other work around would be with the `not` operator,
but that makes queries more verbose and less readeable
See https://stackoverflow.com/questions/48601618/strict-inequality-in-css-media-queries