display the collection if the selected facet is this collection
to allow the user to actually click on the collection if they would like
to see more information about it.

I considered disabling the collection click, but it gives a too new interface (nowhere in the app an entity is disabled/opacitied) and an immediate "why cant i just click on it" feeling came in mind

Codewise, 'wdt:P195' is hardcoded as dynamic generation might be YAGNI.
ie. author layout, which could be concern by the behavior, dont currently
show a series facet selector.