work browser: limiting amount of subentities
ie. layout /wdt:P1433-wd:Q180445

Limiting to 200 seems low but this is conservative choice that can be
increased later.
Reason is that it might make some browsers crash (typically on not so recent
smartphone) and that even users with more powerful machines will have to wait
a lot, especially to fetch articles items, which are less likely to be cached
by the server.

for people wanting facets anyway (typically a researcher would mind waiting
for 2 hours, if necessary), there could  be a link with a query parameter
(ie. `&allFacets=true`) in a warning to reload the page with facets.