task controls: add a swap & merge possibility
leave decision to the user to invert fromId and toId
to reduce the amount of documents that need to be updated.
This could be done automatically server side on day.

Having a button is the easiest, as from and to variables dependencies
depends of entity layouts constants
(this is a lesson on why making reactive variables are preferable)