remove tasks backlog
which was designed to solve hundred of bot generated tasks,
but this PR now prioritise small number of user generated tasks.

Removing the backlog simplifies the code, but it comes with a trade off:
if several users do tasks at the same time, the offset parameter will
leave some undone tasks unseen, since getTask request will only return non
archived tasks.

A way to solve this problem would be to make an
additional request before incrementing the offset to, see is it
returned the same task, is not, it means someone else solved a task in
the meanwhile. But probable YAGNI should let us solves this problems
later if problems arise.